Division, Sqrt, request response de division y sqrt, sin validacion server-side

+++ CalculatorService/CalculatorService.ServerAPI/Models/DivisionRequest.cs
@@ -2,7 +2,7 @@
 {
 	public class DivisionRequest
 	{
-		public double Dividend { get; set; }
-		public double Divisor { get; set; }
+		public int Dividend { get; set; }
+		public int Divisor { get; set; }
 	}
 }
