+++ CalculatorService/CalculatorService.ServerAPI/Models/DivisionResponse.cs
@@ -2,10 +2,10 @@
 {
 	public class DivisionResponse
 	{
-		public double? Quotient { get; set; }
-		public double? Remainder { get; set; }
+		public int? Quotient { get; set; }
+		public int? Remainder { get; set; }
 
-		public static DivisionResponse FromDivision(double quotient, double remainder)
+		public static DivisionResponse FromDivision(int quotient, int remainder)
 		{
 			return new DivisionResponse { Quotient = quotient, Remainder = remainder };
 		}
